IMP: moving eggnog stuff from own submodule to FD.
