+++ q2_types_genomics/__init__.py
@@ -15,4 +15,4 @@ del get_versions
 
 importlib.import_module('q2_types_genomics.per_sample_data')
 importlib.import_module('q2_types_genomics.genome_data')
-importlib.import_module('q2_types_genomics.eggnog')
+importlib.import_module('q2_types_genomics.feature_data')
