+++ q2_types_genomics/feature_data/_transformer.py
@@ -15,8 +15,9 @@ import pandas as pd
 import skbio
 from q2_types.feature_data._transformer import _fastaformats_to_series
 
-from . import MAGSequencesDirFmt
+from . import MAGSequencesDirFmt, ArbitraryHeaderTSVFmt
 from ..plugin_setup import plugin
+from qiime2 import Metadata
 
 CONSTRUCTORS = {
     'DNA': skbio.DNA,
@@ -89,3 +90,32 @@ def _5(data: MAGIterator) -> MAGSequencesDirFmt:
         with open(fp, 'a') as fin:
             skbio.io.write(seq, format='fasta', into=fin)
     return result
+
+# EggnogTransformers
+def _parse_annotation_fmt_to_dataframe(ff):
+    df = pd.read_csv(str(ff), sep='\t', header=ff.header)
+    df.columns = [each.strip("#") for each in df.columns]
+    return df
+
+
+def _tsvfmt_to_metadata(data: ArbitraryHeaderTSVFmt) -> Metadata:
+    df = _parse_annotation_fmt_to_dataframe(data)
+    df.index = df.index.astype(str).rename("id")
+    return Metadata(df)
+
+
+@plugin.register_transformer
+def _6(data: ArbitraryHeaderTSVFmt) -> pd.DataFrame:
+    return _parse_annotation_fmt_to_dataframe(data)
+
+
+@plugin.register_transformer
+def _7(df: pd.DataFrame) -> ArbitraryHeaderTSVFmt:
+    ff = ArbitraryHeaderTSVFmt()
+    df.to_csv(str(ff), sep='\t', header=True, index=True)
+    return ff
+
+
+@plugin.register_transformer
+def _8(data: ArbitraryHeaderTSVFmt) -> Metadata:
+    _tsvfmt_to_metadata(data)
