+++ q2_types_genomics/feature_data/tests/data/kegg_annotations.txt
