+++ q2_types_genomics/feature_data/tests/data/nog_annnotations.txt
