+++ q2_types_genomics/feature_data/tests/data/sample_annotations.txt
