+++ q2_types_genomics/plugin_setup.py
@@ -29,4 +29,3 @@ plugin = qiime2.plugin.Plugin(
 importlib.import_module('q2_types_genomics.feature_data')
 importlib.import_module('q2_types_genomics.per_sample_data')
 importlib.import_module('q2_types_genomics.genome_data')
-importlib.import_module('q2_types_genomics.eggnog')
